+++ ReleaseNotes.markdown
@@ -83,6 +83,9 @@ https://github.com/privcount/privcount/milestone/9?closed=1
   Likely triggered by #361 in 1.0.2
 * Handle tor relays that haven't bootstrapped yet #364
   Bugfix on #361 in 1.0.2
+* Distinguish between a missing nickname and an unknown nickname #366
+  Bugfix on 1.0.0
+* Always check if EnablePrivCount is on before starting a collection #365
 
 ## PrivCount 1.0.2